requirements.txt
asgiref==3.3.1
astroid==2.5.2
autopep8==1.5.5
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.7
django-environ==0.4.5
django-heroku==0.3.1
dnspython==2.1.0
flake8==3.8.4
gunicorn==20.1.0
idna==2.10
importlib-metadata==3.10.0
isort==5.8.0
jedi==0.17.2
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
parso==0.7.1
pluggy==0.13.1
psycopg2==2.8.6
pycodestyle==2.6.0
pydocstyle==6.0.0
pyflakes==2.2.0
pylint==2.7.4
pyls-mypy @ git+https://github.com/tomv564/pyls-mypy.git@3b105b491a42eec304f19e6ed6da1053d12958c6
pymongo==3.11.3
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytz==2021.1
requests==2.25.1
rope==0.18.0
snowballstemmer==2.1.0
sqlparse==0.2.4
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
ujson==4.0.2
urllib3==1.26.4
whitenoise==5.2.0
wrapt==1.12.1
yapf==0.31.0
zipp==3.4.1
